Move remaining Enigma power to SeekerMod
