+++ src/main/java/fruitymod/patches/com/megacrit/cards/AbstractCard/CanUseDazed.java
@@ -13,7 +13,7 @@ public class CanUseDazed {
 			public void edit(MethodCall m) throws CannotCompileException {
 				if (m.getMethodName().equals("hasRelic")) {
 					// pass the original argument (relicID) + this ($0 which is the AbstractCard)
-					m.replace("{ $_ = fruitymod.FruityMod.hasRelicCustom($1, this); }");
+					m.replace("{ $_ = fruitymod.SeekerMod.hasRelicCustom($1, this); }");
 				}
 			}
 		};
